+++ dart/dynamics/BodyNode.hpp
@@ -39,6 +39,7 @@
 #include <Eigen/StdVector>
 
 #include "dart/config.hpp"
+#include "dart/common/Deprecated.hpp"
 #include "dart/common/Signal.hpp"
 #include "dart/common/EmbeddedAspect.hpp"
 #include "dart/math/Geometry.hpp"
@@ -802,11 +803,22 @@ public:
   // Energies
   //----------------------------------------------------------------------------
 
+  /// Return Lagrangian of this body
+  double computeLagrangian(const Eigen::Vector3d& gravity) const;
+
   /// Return kinetic energy.
+  DART_DEPRECATED(6.1)
   virtual double getKineticEnergy() const;
 
+  /// Return kinetic energy
+  double computeKineticEnergy() const;
+
   /// Return potential energy.
+  DART_DEPRECATED(6.1)
   virtual double getPotentialEnergy(const Eigen::Vector3d& _gravity) const;
+
+  /// Return potential energy.
+  double computePotentialEnergy(const Eigen::Vector3d& gravity) const;
 
   /// Return linear momentum.
   Eigen::Vector3d getLinearMomentum() const;